backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pyjwt==2.8.0
redis==5.0.1
slowapi==0.1.9
cryptography>=41.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.1
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3 